\name{add_assembly}
\alias{add_assembly}
\title{add new assembly to copynumber pakcage.}
\description{
  add new assembly to copynumber pakcage.
}
\usage{
add_assembly(name)
}
\arguments{
  \item{name}{a scalar string, indicates the genome assembly cytoBand added into copynumber package. Notes: should be the same with the object name as \code{copynumber} package use \code{get} to derive this object.}
}

\value{
  a character vector of \code{supported_assembly} invisibly
}

