% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logR}
\alias{logR}
\alias{BAF}
\title{Artificial SNP array data}
\format{
Two corresponding data sets containing 10000 probes with logR
and BAF measurements, respectively, for 2 samples. The two first
columns in both data sets contain chromosome numbers and local probe
positions (in base pairs), while the subsequent columns contain
logR-values and BAF-values in the two data sets, respectively.
}
\usage{
data(logR)
data(BAF)
}
\description{
Artificial SNP array data containing a logR track and a BAF track
}
\examples{
#Get data
data(logR)
data(BAF)
}
\keyword{datasets}
