% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber-package.r
\docType{package}
\name{copynumber-package}
\alias{copynumber}
\alias{copynumber-package}
\title{copynumber: Segmentation of Single- and Multi-Track Copy Number Data by Penalized Least Squares Regression (with hg38/mm10/mm39)}
\description{
Penalized least squares regression is applied to fit piecewise constant curves to copy number data to locate genomic regions of constant copy number. Procedures are available for individual segmentation of each sample, joint segmentation of several samples and joint segmentation of the two data tracks from SNP-arrays. Several plotting functions are available for visualization of the data and the segmentation results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://igordot.github.io/copynumber/}
  \item Report bugs at \url{https://github.com/igordot/copynumber/issues}
}

}
\author{
\strong{Maintainer}: Gro Nilsen \email{gronilse@ifi.uio.no}

Authors:
\itemize{
  \item Gro Nilsen \email{gronilse@ifi.uio.no}
  \item Knut Liestol
  \item Ole Christian Lingjaerde
  \item Igor Dolgalev \email{igor.dolgalev@nyumc.org} (\href{https://orcid.org/0000-0003-4451-126X}{ORCID})
}

}
